# docs/swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trigger/{station_id}": {
      "post": {
        "description": "Triggers completion of current task at station and schedules next task. Station must be OCCUPIED (not RESERVED).",
        "operationId": "TriggerController_triggerStation",
        "parameters": [
          {
            "name": "station_id",
            "required": true,
            "in": "path",
            "description": "The ID of the station to trigger",
            "schema": {
              "example": "ST001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Trigger station action payload",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dropped_quantity": {
                    "type": "integer",
                    "example": 5
                  },
                  "message_code": {
                    "type": "string",
                    "enum": [
                      "SEND_TO_EMPTY_LOCATION"
                    ],
                    "example": "NO_PROBLEM"
                  }
                },
                "required": [
                  "dropped_quantity",
                  "message_code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Station triggered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Station ST001 triggered successfully"
                        },
                        "triggered_task": {
                          "type": "object"
                        },
                        "station": {
                          "type": "object"
                        },
                        "timestamp": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Station ST001 triggered successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to trigger station",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Failed to trigger station"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Station cannot be triggered (reserved or not occupied)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Can't trigger now - station ST001 is reserved"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Trigger station action",
        "tags": [
          "Trigger"
        ]
      }
    },
...
  },
  "info": {
    "title": "ARAPL RaaS GTP API",
    "description": "\n      Goods To Person (GTP) API for ARAPL RaaS (Robotics as a Service)\n\n      This API provides comprehensive GTP capabilities including:\n      - Order management and file upload processing\n      - Moveops location management\n      - Station management and Pallet Slot assignments\n      - Inventory tracking and bulk operations\n      \n      All endpoints include proper validation, error handling, and support file uploads where applicable.\n    ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "API Key for authentication"
      }
    },
...
  }
}